Automatically build image
With this change there is no longer any need to manually create an ECR repo, and manually build and push your Docker image. Instead CDK will automatically build the image using the Dockerfile in the "api" directory, automatically create an ECR repo and push the image, and then use it in the Fargate service.